+++ Braintree.Tests/TransactionTest.cs
@@ -1215,7 +1215,7 @@ namespace Braintree.Tests
             var request = new TransactionRequest
             {
                 Amount = SandboxValues.TransactionAmount.AUTHORIZE,
-                MerchantAccountId = MerchantAccount.NON_DEFAULT_MERCHANT_ACCOUNT_ID,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_MERCHANT_ACCOUNT_ID,
                 CreditCard = new TransactionCreditCardRequest
                 {
                     Number = SandboxValues.CreditCardNumber.VISA,
@@ -1227,7 +1227,7 @@ namespace Braintree.Tests
             Assert.IsTrue(result.IsSuccess());
             Transaction transaction = result.Target;
 
-            Assert.AreEqual(MerchantAccount.NON_DEFAULT_MERCHANT_ACCOUNT_ID, transaction.MerchantAccountId);
+            Assert.AreEqual(MerchantAccountIDs.NON_DEFAULT_MERCHANT_ACCOUNT_ID, transaction.MerchantAccountId);
         }
 
         [Test]
@@ -1247,7 +1247,7 @@ namespace Braintree.Tests
             Assert.IsTrue(result.IsSuccess());
             Transaction transaction = result.Target;
 
-            Assert.AreEqual(MerchantAccount.DEFAULT_MERCHANT_ACCOUNT_ID, transaction.MerchantAccountId);
+            Assert.AreEqual(MerchantAccountIDs.DEFAULT_MERCHANT_ACCOUNT_ID, transaction.MerchantAccountId);
         }
 
         [Test]
@@ -1906,6 +1906,307 @@ namespace Braintree.Tests
         }
 
         [Test]
+        public void Sale_WithServiceFee()
+        {
+            var request = new TransactionRequest
+            {
+                Amount = SandboxValues.TransactionAmount.AUTHORIZE,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_SUB_MERCHANT_ACCOUNT_ID,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+                ServiceFeeAmount = 1M
+            };
+
+            Result<Transaction> result = gateway.Transaction.Sale(request);
+            Assert.IsTrue(result.IsSuccess());
+
+            Transaction transaction = result.Target;
+            Assert.AreEqual(1M, transaction.ServiceFeeAmount);
+        }
+
+        [Test]
+        public void Sale_WithServiceFeeWithTooLargeAmount()
+        {
+            TransactionRequest request = new TransactionRequest
+            {
+                Amount = 1M,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_SUB_MERCHANT_ACCOUNT_ID,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+                ServiceFeeAmount = 2M
+            };
+
+            Result<Transaction> result = gateway.Transaction.Sale(request);
+            Assert.IsFalse(result.IsSuccess());
+            Assert.AreEqual(ValidationErrorCode.TRANSACTION_SERVICE_FEE_AMOUNT_IS_TOO_LARGE, result.Errors.ForObject("Transaction").OnField("ServiceFeeAmount")[0].Code);
+        }
+
+        [Test]
+        public void Sale_WithMerchantAccountIdAndWithoutServiceFeeAmount()
+        {
+            TransactionRequest request = new TransactionRequest
+            {
+                Amount = 1M,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_SUB_MERCHANT_ACCOUNT_ID,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                }
+            };
+
+            Result<Transaction> result = gateway.Transaction.Sale(request);
+            Assert.IsFalse(result.IsSuccess());
+            Assert.AreEqual(ValidationErrorCode.TRANSACTION_SUB_MERCHANT_ACCOUNT_REQUIRES_SERVICE_FEE_AMOUNT, result.Errors.ForObject("Transaction").OnField("MerchantAccountId")[0].Code);
+        }
+
+        [Test]
+        public void Sale_WithServiceFeeAmountOnMasterMerchantAccount()
+        {
+            TransactionRequest request = new TransactionRequest
+            {
+                Amount = 4M,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+                ServiceFeeAmount = 2M
+            };
+
+            Result<Transaction> result = gateway.Transaction.Sale(request);
+            Assert.IsFalse(result.IsSuccess());
+            Assert.AreEqual(ValidationErrorCode.TRANSACTION_SERVICE_FEE_AMOUNT_NOT_ALLOWED_ON_MASTER_MERCHANT_ACCOUNT , result.Errors.ForObject("Transaction").OnField("ServiceFeeAmount")[0].Code);
+        }
+
+        [Test]
+        public void Sale_WithHoldInEscrow()
+        {
+            TransactionRequest request = new TransactionRequest
+            {
+                Amount = 4M,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_SUB_MERCHANT_ACCOUNT_ID,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+                ServiceFeeAmount = 2M,
+                Options = new TransactionOptionsRequest
+                {
+                    HoldInEscrow = true
+                }
+            };
+            Result<Transaction> result = gateway.Transaction.Sale(request);
+            Assert.IsTrue(result.IsSuccess());
+            Transaction transaction = result.Target;
+            Assert.AreEqual(TransactionEscrowStatus.HOLD_PENDING, transaction.EscrowStatus);
+        }
+
+        [Test]
+        public void Sale_WithHoldInEscrowFailsForMasterMerchantAccount()
+        {
+            TransactionRequest request = new TransactionRequest
+            {
+                Amount = 4M,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_MERCHANT_ACCOUNT_ID,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+                Options = new TransactionOptionsRequest
+                {
+                    HoldInEscrow = true
+                }
+            };
+            Result<Transaction> result = gateway.Transaction.Sale(request);
+            Assert.IsFalse(result.IsSuccess());
+            Assert.AreEqual(
+                ValidationErrorCode.TRANSACTION_CANNOT_HOLD_IN_ESCROW,
+                result.Errors.ForObject("Transaction").OnField("Base")[0].Code
+            );
+        }
+
+        [Test]
+        public void HoldInEscrow_AfterSale()
+        {
+            TransactionRequest request = new TransactionRequest
+            {
+                Amount = 4M,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_SUB_MERCHANT_ACCOUNT_ID,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+                ServiceFeeAmount = 1M
+            };
+            Result<Transaction> saleResult = gateway.Transaction.Sale(request);
+            Transaction saleTransaction = saleResult.Target;
+            Result<Transaction> result = gateway.Transaction.HoldInEscrow(saleTransaction.Id);
+            Assert.IsTrue(result.IsSuccess());
+            Transaction transaction = result.Target;
+            Assert.AreEqual(
+                TransactionEscrowStatus.HOLD_PENDING,
+                transaction.EscrowStatus
+            );
+        }
+
+        [Test]
+        public void HoldInEscrow_AfterSaleFailsForMasterMerchantAccount()
+        {
+            TransactionRequest request = new TransactionRequest
+            {
+                Amount = 4M,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_MERCHANT_ACCOUNT_ID,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+            };
+            Result<Transaction> saleResult = gateway.Transaction.Sale(request);
+            Transaction saleTransaction = saleResult.Target;
+            Result<Transaction> result = gateway.Transaction.HoldInEscrow(saleTransaction.Id);
+            Assert.IsFalse(result.IsSuccess());
+            Assert.AreEqual(
+                ValidationErrorCode.TRANSACTION_CANNOT_HOLD_IN_ESCROW,
+                result.Errors.ForObject("Transaction").OnField("Base")[0].Code
+            );
+        }
+
+        [Test]
+        public void ReleaseFromEscrow()
+        {
+            TransactionRequest request = new TransactionRequest
+            {
+                Amount = 4M,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_SUB_MERCHANT_ACCOUNT_ID,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+                Options = new TransactionOptionsRequest
+                {
+                    HoldInEscrow = true
+                },
+                ServiceFeeAmount = 1M
+            };
+            Result<Transaction> saleResult = gateway.Transaction.Sale(request);
+            Transaction saleTransaction = saleResult.Target;
+            Assert.IsTrue(saleResult.IsSuccess());
+            TestHelper.Escrow(service, saleTransaction.Id);
+            Result<Transaction> result = gateway.Transaction.ReleaseFromEscrow(saleTransaction.Id);
+
+            Assert.IsTrue(result.IsSuccess());
+            Transaction transaction = result.Target;
+            Assert.AreEqual(
+                TransactionEscrowStatus.RELEASE_PENDING,
+                transaction.EscrowStatus
+            );
+        }
+
+        [Test]
+        public void ReleaseFromEscrow_FailsForNonSubmittableTransaction()
+        {
+            TransactionRequest request = new TransactionRequest
+            {
+                Amount = 4M,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_MERCHANT_ACCOUNT_ID,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                }
+            };
+            Result<Transaction> saleResult = gateway.Transaction.Sale(request);
+            Transaction saleTransaction = saleResult.Target;
+
+            Result<Transaction> result = gateway.Transaction.ReleaseFromEscrow(saleTransaction.Id);
+
+            Assert.IsFalse(result.IsSuccess());
+            Assert.AreEqual(
+                ValidationErrorCode.TRANSACTION_CANNOT_RELEASE_FROM_ESCROW,
+                result.Errors.ForObject("Transaction").OnField("Base")[0].Code
+            );
+        }
+
+        [Test]
+        public void CancelRelease()
+        {
+            TransactionRequest request = new TransactionRequest
+            {
+                Amount = 4M,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_SUB_MERCHANT_ACCOUNT_ID,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+                Options = new TransactionOptionsRequest
+                {
+                    HoldInEscrow = true
+                },
+                ServiceFeeAmount = 1M
+            };
+            Result<Transaction> saleResult = gateway.Transaction.Sale(request);
+            Transaction saleTransaction = saleResult.Target;
+            Assert.IsTrue(saleResult.IsSuccess());
+            TestHelper.Escrow(service, saleTransaction.Id);
+            Result<Transaction> result = gateway.Transaction.ReleaseFromEscrow(saleTransaction.Id);
+            Assert.IsTrue(result.IsSuccess());
+
+            Transaction releasedTransaction = result.Target;
+
+            Result<Transaction> cancelResult = gateway.Transaction.CancelRelease(releasedTransaction.Id);
+            Assert.IsTrue(cancelResult.IsSuccess());
+            Transaction transaction = cancelResult.Target;
+            Assert.AreEqual(
+                TransactionEscrowStatus.HELD,
+                transaction.EscrowStatus
+            );
+        }
+
+        [Test]
+        public void CancelRelease_FailsForTransactionsNotPendingRelease()
+        {
+            TransactionRequest request = new TransactionRequest
+            {
+                Amount = 4M,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_SUB_MERCHANT_ACCOUNT_ID,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+                Options = new TransactionOptionsRequest
+                {
+                    HoldInEscrow = true
+                },
+                ServiceFeeAmount = 1M
+            };
+            Result<Transaction> saleResult = gateway.Transaction.Sale(request);
+            Transaction saleTransaction = saleResult.Target;
+            Assert.IsTrue(saleResult.IsSuccess());
+            TestHelper.Escrow(service, saleTransaction.Id);
+
+            Result<Transaction> cancelResult = gateway.Transaction.CancelRelease(saleTransaction.Id);
+            Assert.IsFalse(cancelResult.IsSuccess());
+            Assert.AreEqual(
+                ValidationErrorCode.TRANSACTION_CANNOT_CANCEL_RELEASE,
+                cancelResult.Errors.ForObject("Transaction").OnField("Base")[0].Code
+            );
+        }
+
+        [Test]
         public void Sale_WithDescriptor()
         {
             var request = new TransactionRequest
@@ -2011,6 +2312,30 @@ namespace Braintree.Tests
             Assert.AreEqual("1111", transaction.CreditCard.LastFour);
         }
 
+        [Test]
+        public void Sale_WithVenmoSdkSession()
+        {
+            var request = new TransactionRequest
+            {
+                Amount = SandboxValues.TransactionAmount.AUTHORIZE,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+                Options = new TransactionOptionsRequest
+                {
+                    VenmoSdkSession = SandboxValues.VenmoSdk.SESSION
+                }
+            };
+
+            Result<Transaction> result = gateway.Transaction.Sale(request);
+            Assert.IsTrue(result.IsSuccess());
+
+            Transaction transaction = result.Target;
+            Assert.IsTrue(transaction.CreditCard.IsVenmoSdk.Value);
+        }
+
         #pragma warning disable 0618
         [Test]
         public void ConfirmTransparentRedirect_CreatesTheTransaction()
@@ -2071,7 +2396,7 @@ namespace Braintree.Tests
             {
                 Amount = SandboxValues.TransactionAmount.AUTHORIZE,
                 Type = TransactionType.SALE,
-                MerchantAccountId = MerchantAccount.NON_DEFAULT_MERCHANT_ACCOUNT_ID
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_MERCHANT_ACCOUNT_ID
             };
 
             TransactionRequest request = new TransactionRequest
@@ -2088,7 +2413,7 @@ namespace Braintree.Tests
             Assert.IsTrue(result.IsSuccess());
             Transaction transaction = result.Target;
 
-            Assert.AreEqual(MerchantAccount.NON_DEFAULT_MERCHANT_ACCOUNT_ID, transaction.MerchantAccountId);
+            Assert.AreEqual(MerchantAccountIDs.NON_DEFAULT_MERCHANT_ACCOUNT_ID, transaction.MerchantAccountId);
         }
         #pragma warning restore 0618
 
@@ -2127,7 +2452,7 @@ namespace Braintree.Tests
             var request = new TransactionRequest
             {
                 Amount = SandboxValues.TransactionAmount.AUTHORIZE,
-                MerchantAccountId = MerchantAccount.NON_DEFAULT_MERCHANT_ACCOUNT_ID,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_MERCHANT_ACCOUNT_ID,
                 CreditCard = new TransactionCreditCardRequest
                 {
                     Number = SandboxValues.CreditCardNumber.VISA,
@@ -2139,7 +2464,7 @@ namespace Braintree.Tests
             Assert.IsTrue(result.IsSuccess());
             Transaction transaction = result.Target;
 
-            Assert.AreEqual(MerchantAccount.NON_DEFAULT_MERCHANT_ACCOUNT_ID, transaction.MerchantAccountId);
+            Assert.AreEqual(MerchantAccountIDs.NON_DEFAULT_MERCHANT_ACCOUNT_ID, transaction.MerchantAccountId);
         }
 
         [Test]
@@ -2159,7 +2484,7 @@ namespace Braintree.Tests
             Assert.IsTrue(result.IsSuccess());
             Transaction transaction = result.Target;
 
-            Assert.AreEqual(MerchantAccount.DEFAULT_MERCHANT_ACCOUNT_ID, transaction.MerchantAccountId);
+            Assert.AreEqual(MerchantAccountIDs.DEFAULT_MERCHANT_ACCOUNT_ID, transaction.MerchantAccountId);
         }
 
         [Test]
@@ -2210,6 +2535,29 @@ namespace Braintree.Tests
             Assert.IsFalse(parameters.ContainsKey("transaction[amount]"));
             Assert.AreEqual("05", parameters["transaction[credit_card][expiration_month]"]);
             Assert.AreEqual("2010", parameters["transaction[credit_card][expiration_year]"]);
+        }
+
+        [Test]
+        public void Credit_WithServiceFeeIsDisallowed()
+        {
+            var request = new TransactionRequest
+            {
+                Amount = SandboxValues.TransactionAmount.AUTHORIZE,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                },
+                ServiceFeeAmount = 1M
+            };
+
+            Result<Transaction> result = gateway.Transaction.Credit(request);
+            Assert.IsFalse(result.IsSuccess());
+
+            Assert.AreEqual(
+                ValidationErrorCode.TRANSACTION_SERVICE_FEE_IS_NOT_ALLOWED_ON_CREDITS,
+                result.Errors.ForObject("Transaction").OnField("Base")[0].Code
+            );
         }
 
         [Test]
@@ -2390,6 +2738,31 @@ namespace Braintree.Tests
             Assert.IsTrue(result.IsSuccess());
             Assert.AreEqual(TransactionStatus.SUBMITTED_FOR_SETTLEMENT, result.Target.Status);
             Assert.AreEqual(50.00, result.Target.Amount);
+        }
+
+        [Test]
+        public void SubmitForSettlement_TransactionAmountLessThanServiceFeeAmount()
+        {
+            TransactionRequest request = new TransactionRequest
+            {
+                Amount = 100M,
+                MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_SUB_MERCHANT_ACCOUNT_ID,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "06/2008"
+                },
+                ServiceFeeAmount = 50M
+            };
+
+            Transaction transaction = gateway.Transaction.Sale(request).Target;
+            Result<Transaction> result = gateway.Transaction.SubmitForSettlement(transaction.Id, 25M);
+
+            Assert.IsFalse(result.IsSuccess());
+            Assert.AreEqual(
+                ValidationErrorCode.TRANSACTION_SETTLEMENT_AMOUNT_IS_LESS_THAN_SERVICE_FEE_AMOUNT,
+                result.Errors.ForObject("Transaction").OnField("Amount")[0].Code
+            );
         }
 
         [Test]