+++ Braintree.Tests/WebhookNotificationTest.cs
@@ -59,5 +59,46 @@ namespace Braintree.Tests
             Dictionary<String, String> sampleNotification = gateway.WebhookTesting.SampleNotification(WebhookKind.SUBSCRIPTION_WENT_PAST_DUE, "my_id");
             gateway.WebhookNotification.Parse("bad" + sampleNotification["signature"], sampleNotification["payload"]);
         }
+
+        [Test]
+        public void SampleNotification_ReturnsANotificationForAMerchantAccountApprovedWebhook()
+        {
+            Dictionary<String, String> sampleNotification = gateway.WebhookTesting.SampleNotification(WebhookKind.SUB_MERCHANT_ACCOUNT_APPROVED, "my_id");
+
+            WebhookNotification notification = gateway.WebhookNotification.Parse(sampleNotification["signature"], sampleNotification["payload"]);
+
+            Assert.AreEqual(WebhookKind.SUB_MERCHANT_ACCOUNT_APPROVED, notification.Kind);
+            Assert.AreEqual("my_id", notification.MerchantAccount.Id);
+            Assert.AreEqual(MerchantAccountStatus.ACTIVE, notification.MerchantAccount.Status);
+            Assert.AreEqual("master_ma_for_my_id", notification.MerchantAccount.MasterMerchantAccount.Id);
+            Assert.AreEqual(MerchantAccountStatus.ACTIVE, notification.MerchantAccount.MasterMerchantAccount.Status);
+            TestHelper.AreDatesEqual(DateTime.Now.ToUniversalTime(), notification.Timestamp.Value);
+        }
+
+        [Test]
+        public void SampleNotification_ReturnsANotificationForAMerchantAccountDeclinedWebhook()
+        {
+          Dictionary<String, String> sampleNotification = gateway.WebhookTesting.SampleNotification(WebhookKind.SUB_MERCHANT_ACCOUNT_DECLINED, "my_id");
+
+          WebhookNotification notification = gateway.WebhookNotification.Parse(sampleNotification["signature"], sampleNotification["payload"]);
+
+          Assert.AreEqual(WebhookKind.SUB_MERCHANT_ACCOUNT_DECLINED, notification.Kind);
+          Assert.AreEqual("my_id", notification.MerchantAccount.Id);
+          Assert.AreEqual(ValidationErrorCode.MERCHANT_ACCOUNT_APPLICANT_DETAILS_DECLINED_OFAC, notification.Errors.ForObject("merchant-account").OnField("base")[0].Code);
+          Assert.AreEqual("Applicant declined due to OFAC.", notification.Message);
+        }
+
+        [Test]
+        public void SampleNotification_ReturnsANotificationForATransactionDisbursedWebhook()
+        {
+          Dictionary<String, String> sampleNotification = gateway.WebhookTesting.SampleNotification(WebhookKind.TRANSACTION_DISBURSED, "my_id");
+
+          WebhookNotification notification = gateway.WebhookNotification.Parse(sampleNotification["signature"], sampleNotification["payload"]);
+
+          Assert.AreEqual(WebhookKind.TRANSACTION_DISBURSED, notification.Kind);
+          Assert.AreEqual(100, notification.Transaction.Amount);
+          Assert.AreEqual("my_id", notification.Transaction.Id);
+          Assert.IsTrue(notification.Transaction.DisbursementDetails.IsValid());
+        }
     }
 }