+++ Braintree/MerchantAccount.cs
@@ -1,0 +1,39 @@
+using System;
+
+namespace Braintree
+{
+    public class MerchantAccountStatus : Enumeration
+    {
+        public static readonly MerchantAccountStatus PENDING = new MerchantAccountStatus("pending");
+        public static readonly MerchantAccountStatus ACTIVE = new MerchantAccountStatus("active");
+        public static readonly MerchantAccountStatus SUSPENDED = new MerchantAccountStatus("suspended");
+
+        public static readonly MerchantAccountStatus[] ALL = { PENDING, ACTIVE, SUSPENDED };
+
+        protected MerchantAccountStatus(String name) : base(name) {}
+    }
+
+    public class MerchantAccount
+    {
+      public String Id { get; protected set; }
+      public MerchantAccountStatus Status { get; protected set; }
+      public MerchantAccount MasterMerchantAccount { get; protected set; }
+
+      public Boolean IsSubMerchant {
+        get {
+          return MasterMerchantAccount != null;
+        }
+      }
+
+      protected internal MerchantAccount(NodeWrapper node)
+      {
+        Id = node.GetString("id");
+        Status = (MerchantAccountStatus) CollectionUtil.Find(MerchantAccountStatus.ALL, node.GetString("status"), null);
+        NodeWrapper masterNode = node.GetNode("master-merchant-account");
+        if (masterNode != null)
+            MasterMerchantAccount = new MerchantAccount(masterNode);
+        else
+            MasterMerchantAccount = null;
+      }
+    }
+}