+++ Braintree/Test/MerchantAccount.cs
@@ -1,0 +1,10 @@
+namespace Braintree
+{
+    namespace Test
+    {
+        public class MerchantAccount
+        {
+            public static String Appove = "approve_me";
+        }
+    }
+}